Generalize bluenop as much as possible, add to blueno package
