+++ blueno/__init__.py
@@ -15,3 +15,4 @@ from .utils import (
 )
 
 from .pipeline.bluenot import start_train
+from .pipeline.bluenop import start_preprocess
